requirements.txt
cohere==5.20.1
discord.py==2.6.4
pytz==2025.2
requests==2.32.5
python-dotenv==1.2.1
